Refactor modules to top-level & add guard to Unifrac
