+++ onecodex/distance.py
@@ -1,6 +1,6 @@
 import skbio.diversity
 
-from onecodex.lib.taxonomy import generate_skbio_tree, prune_to_rank
+from onecodex.taxonomy import generate_skbio_tree, prune_to_rank
 
 
 ACCEPTABLE_FIELDS = ['abundance', 'readcount_w_children', 'readcount']
@@ -16,7 +16,7 @@ def alpha_counts(classification, field='readcount_w_children', rank='species'):
 
 
 def beta_counts(classifications, field='readcount_w_children', rank='species'):
-    from onecodex.viz.helpers import normalize_analyses, collate_analysis_results
+    from onecodex.helpers import normalize_analyses, collate_analysis_results
     normed_analyses, _ = normalize_analyses(classifications)
     df = collate_analysis_results(normed_analyses, field=field)
     df = df.loc[:, [i[2] == rank for i in df.columns]]
@@ -73,17 +73,19 @@ def unifrac(classifications, weighted=True,
     A beta diversity metric that takes into account the relative relatedness of community members.
     Weighted UniFrac looks at abundances, unweighted UniFrac looks at presence
     """
-    assert field in ACCEPTABLE_FIELDS
+    from skbio.tree import TreeNode
 
+    assert field in ACCEPTABLE_FIELDS
     counts, tax_ids, ids = beta_counts(classifications, field=field, rank=rank)
 
     tree = None
     for c in classifications:
+        assert c.job.id == classifications[0].job.id, "All classifications must " \
+                                                      "have same job for Unifrac"
         tree = generate_skbio_tree(c, existing_tree=tree)
 
     # there's a bug (?) in skbio where it expects the root to only have
     # one child, so we do a little faking here
-    from skbio.tree import TreeNode
     new_tree = TreeNode(name='fake root')
     new_tree.rank = 'no rank'
     new_tree.append(tree)
