+++ onecodex/viz/distance.py
@@ -1,8 +1,8 @@
 import pandas as pd
 
 from onecodex.exceptions import OneCodexException
-from onecodex.viz.helpers import normalize_analyses
-from onecodex.lib.distance_metrics import braycurtis, cityblock, jaccard
+from onecodex.helpers import normalize_analyses
+from onecodex.distance import braycurtis, cityblock, jaccard
 
 
 def plot_distance(analyses, title=None, distance_metric='bray-curtis',
