+++ onecodex/viz/heatmap.py
@@ -1,5 +1,5 @@
 from onecodex.exceptions import OneCodexException
-from onecodex.viz.helpers import collate_analysis_results, normalize_analyses
+from onecodex.helpers import collate_analysis_results, normalize_analyses
 
 
 def plot_heatmap(analyses, title=None, top_n=20, threshold=None, field='readcount_w_children'):
